+++ multistreaming-server/launch-nginx-server.sh
@@ -8,7 +8,7 @@ envsubst < base-nginx.conf >  /usr/local/nginx/conf/nginx.conf
 /usr/bin/stunnel &
 
 # append nginx conf with RTMP Configuration
-python3 /rtmp-conf-generator.py /rtmp-configuation.json >> /usr/local/nginx/conf/nginx.conf
+python3 /rtmp-conf-generator.py /rtmp-configuration.json >> /usr/local/nginx/conf/nginx.conf
 if [ $? -ne 0 ]; then
   echo "ERROR encountered when generating RTMP configuration."
   exit 1